+++ src/core/mqtt.c
@@ -166,6 +166,11 @@ char *mqtt_state_names[kMqttState_Max] =
 };
 
 //------------------------------------------------------------------------------
+// Buffer containing root cause of failing to initiate a send
+// If the message that couldn't be sent was a USP response message, then this string is sent to the Controller in an ERROR response
+static char mqtt_publish_err_msg[128];
+
+//------------------------------------------------------------------------------
 // Payload to send in MQTT queue
 typedef struct
 {
@@ -254,6 +259,7 @@ void QueueUspMqttConnectRecords(mqtt_client_t *client);
 void QueueUspRecord_MQTT(mqtt_client_t *client, mtp_send_item_t *msi, char *controller_topic, time_t expiry_time, bool link_to_head);
 void MqttSubscriptionReplace(mqtt_subscription_t *dest, mqtt_subs_config_t *src);
 void MqttSubscriptionDestroy(mqtt_subscription_t *sub);
+void SaveMqttPublishErrMsg(const char *fmt, ...);
 
 //------------------------------------------------------------------------------------
 #define DEFINE_MQTT_TrustCertVerifyCallbackIndex(index) \
@@ -1223,6 +1229,9 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
 {
     int i;
     int sock;
+    int err;
+    mqtt_send_item_t *q_msg;
+    mqtt_client_t *client;
 
     OS_UTILS_LockMutex(&mqtt_access_mutex);
 
@@ -1237,7 +1246,6 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
     SOCKET_SET_UpdateTimeout(1*SECONDS, set);
 
     // Iterate over all mqtt clients currently enabled
-    mqtt_client_t* client = NULL;
     for (i = 0; i < MAX_MQTT_CLIENTS; i++)
     {
         client = &mqtt_clients[i];
@@ -1249,12 +1257,14 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
                     RemoveExpiredMqttMessages(client);
                     if ((client->usp_record_send_queue.head) && (client->is_subscribed))
                     {
-                        if (SendQueueHead(client) == USP_ERR_OK)
+                        // Initiate send. Note: if an error occured, it will be logged in mqtt_publish_err_msg[]
+                        err = SendQueueHead(client);
+                        if (err == USP_ERR_OK)
                         {
+                            // Initiated send successfully
                             // If this is a USP disconnect record, then save the (libmosquitto allocated) message id, so that
                             // when we get the publish callback, we can then shutdown the connection
                             // NOTE: USP Disconnect records terminating an E2E session (kMtpContentType_E2E_SessTermination) do not shutdown the connection
-                            mqtt_send_item_t *q_msg;
                             q_msg = (mqtt_send_item_t *) client->usp_record_send_queue.head;
                             client->disconnect_mid = (q_msg->item.content_type == kMtpContentType_DisconnectRecord) ? q_msg->mid : INVALID_MOSQUITTO_MID;
 
@@ -1263,7 +1273,16 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
                         }
                         else
                         {
-                            USP_LOG_Error("%s: Failed to send head of the queue, leaving there to try again", __FUNCTION__);
+                            // Failed to initiate send
+                            // Replace the item to send with a USP ERROR or drop the item, to prevent send queue deadlock
+                            q_msg = (mqtt_send_item_t *) client->usp_record_send_queue.head;
+                            if (q_msg != NULL)
+                            {
+                                if (MSG_HANDLER_ShouldDropFailedMessage(&q_msg->item, mqtt_publish_err_msg) == true)
+                                {
+                                    RemoveMqttQueueItem(client, q_msg);
+                                }
+                            }
                         }
                     }
                     break;
@@ -1335,6 +1354,7 @@ void MQTT_ProcessAllSocketActivity(socket_set_t* set)
 {
     int i;
     int sock;
+    int mosq_err;
     mqtt_client_t *client;
 
     OS_UTILS_LockMutex(&mqtt_access_mutex);
@@ -1364,9 +1384,11 @@ void MQTT_ProcessAllSocketActivity(socket_set_t* set)
                         // Send all data on socket - if required
                         if (SOCKET_SET_IsReadyToWrite(sock, set) && mosquitto_want_write(client->mosq))
                         {
-                            if (mosquitto_loop_write(client->mosq, 1) != MOSQ_ERR_SUCCESS)
+                            mosq_err = mosquitto_loop_write(client->mosq, 1);
+                            if (mosq_err != MOSQ_ERR_SUCCESS)
                             {
-                                USP_LOG_Error("%s: Failed to write to socket", __FUNCTION__);
+                                USP_LOG_Error("%s: Failed to write to socket: %s", __FUNCTION__, mosquitto_strerror(mosq_err));
+                                HandleMqttError(client, kMqttFailure_ReadWrite, "write failure");
                             }
                         }
                     }
@@ -1376,9 +1398,11 @@ void MQTT_ProcessAllSocketActivity(socket_set_t* set)
                     {
                         if (SOCKET_SET_IsReadyToRead(sock, set))
                         {
-                            if (mosquitto_loop_read(client->mosq, 1) != MOSQ_ERR_SUCCESS)
+                            mosq_err = mosquitto_loop_read(client->mosq, 1);
+                            if (mosq_err != MOSQ_ERR_SUCCESS)
                             {
-                                USP_LOG_Error("%s: Failed to read from socket", __FUNCTION__);
+                                USP_LOG_Error("%s: Failed to read from socket: %s", __FUNCTION__, mosquitto_strerror(mosq_err));
+                                HandleMqttError(client, kMqttFailure_ReadWrite, "read failure");
                             }
                         }
                     }
@@ -1386,9 +1410,11 @@ void MQTT_ProcessAllSocketActivity(socket_set_t* set)
                     sock = mosquitto_socket(client->mosq);
                     if (sock != -1)   // NOTE: Retesting, as connection may have been closed after reading
                     {
-                        if (mosquitto_loop_misc(client->mosq) != MOSQ_ERR_SUCCESS)
+                        mosq_err = mosquitto_loop_misc(client->mosq);
+                        if (mosq_err != MOSQ_ERR_SUCCESS)
                         {
-                            USP_LOG_Error("%s: Failed to write misc", __FUNCTION__);
+                            USP_LOG_Error("%s: Failed to write misc: %s", __FUNCTION__, mosquitto_strerror(mosq_err));
+                            HandleMqttError(client, kMqttFailure_ReadWrite, "misc write failure");
                         }
                     }
                     break;
@@ -2302,6 +2328,7 @@ void QueueUspRecord_MQTT(mqtt_client_t *client, mtp_send_item_t *msi, char *cont
 ** SendQueueHead
 **
 ** Publishes the USP record at the front of the queue
+** NOTE: If an error occurs it is saved in mqtt_publish_err_msg[]
 **
 ** \param   client - pointer to MQTT client
 **
@@ -2315,6 +2342,9 @@ int SendQueueHead(mqtt_client_t *client)
     // Can't be passed a NULL client
     USP_ASSERT(client != NULL);
 
+    // Save default cause of failure as empty string. If the ACS receives this in an error message, then there is some path which does not save the cause of error
+    mqtt_publish_err_msg[0] = '\0';
+
     mqtt_state_t state = client->state;
     if (state == kMqttState_Running)
     {
@@ -2323,7 +2353,7 @@ int SendQueueHead(mqtt_client_t *client)
         // Check the queue head is ok
         if (q_msg == NULL)
         {
-            USP_LOG_Error("%s: Can't send NULL head", __FUNCTION__);
+            SaveMqttPublishErrMsg("%s: Can't send NULL head", __FUNCTION__);
             return USP_ERR_INTERNAL_ERROR;
         }
 
@@ -2331,7 +2361,7 @@ int SendQueueHead(mqtt_client_t *client)
     }
     else
     {
-        USP_LOG_Error("%s: Incorrect state for sending messages %s", __FUNCTION__, mqtt_state_names[state]);
+        SaveMqttPublishErrMsg("%s: Incorrect state for sending messages %s", __FUNCTION__, mqtt_state_names[state]);
         err = USP_ERR_INTERNAL_ERROR;
     }
 
@@ -3732,6 +3762,7 @@ void UnsubscribeV5Callback(struct mosquitto *mosq, void *userdata, int mid, cons
 ** Publish
 **
 ** Initiates an MQTT PUBLISH for the specified USP record
+** NOTE: If an error occurs it is saved in mqtt_publish_err_msg[]
 **
 ** \param   client - pointer to MQTT client
 ** \param   msg - pointer to message to send
@@ -3743,6 +3774,7 @@ int Publish(mqtt_client_t *client, mqtt_send_item_t *msg)
 {
     int err = USP_ERR_OK;
     char buf[512];
+    int mosq_err;
 
     USP_ASSERT(client != NULL);
     USP_ASSERT(msg != NULL);
@@ -3772,9 +3804,10 @@ int Publish(mqtt_client_t *client, mqtt_send_item_t *msg)
             USP_STRNCPY(topic, msg->topic, sizeof(topic));
         }
 
-        if (mosquitto_publish(client->mosq, &msg->mid, topic, msg->item.pbuf_len, msg->item.pbuf, msg->qos, false /*retain*/) != MOSQ_ERR_SUCCESS)
+        mosq_err = mosquitto_publish(client->mosq, &msg->mid, topic, msg->item.pbuf_len, msg->item.pbuf, msg->qos, false /*retain*/);
+        if (mosq_err != MOSQ_ERR_SUCCESS)
         {
-            USP_LOG_Error("%s: Failed to publish message to topic=%s", __FUNCTION__, topic);
+            SaveMqttPublishErrMsg("%s: Failed to publish with error %d (%s)", __FUNCTION__, mosq_err, mosquitto_strerror(mosq_err));
             err = USP_ERR_INTERNAL_ERROR;
         }
 
@@ -3794,6 +3827,7 @@ int Publish(mqtt_client_t *client, mqtt_send_item_t *msg)
 ** PublishV5
 **
 ** Initiates an MQTTv5 PUBLISH for the specified USP record
+** NOTE: If an error occurs it is saved in mqtt_publish_err_msg[]
 **
 ** \param   client - pointer to MQTT client
 ** \param   msg - pointer to message to send
@@ -3816,7 +3850,7 @@ int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
     content_type = (msg->item.content_type == kMtpContentType_BulkDataReport) ? MQTT_BULK_DATA_CONTENT_TYPE : MQTT_USP_CONTENT_TYPE;
     if (mosquitto_property_add_string(&proplist, CONTENT_TYPE, content_type) != MOSQ_ERR_SUCCESS)
     {
-        USP_LOG_Error("%s: Failed to add content type string", __FUNCTION__);
+        SaveMqttPublishErrMsg("%s: Failed to add content type string", __FUNCTION__);
         err = USP_ERR_INTERNAL_ERROR;
         goto error;
     }
@@ -3828,7 +3862,7 @@ int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
         USP_ASSERT((client->response_subscription.topic!= NULL) && (client->response_subscription.topic[0] != '\0')); // SubscribeToAll() should have prevented the code getting here
         if (mosquitto_property_add_string(&proplist, RESPONSE_TOPIC, client->response_subscription.topic) != MOSQ_ERR_SUCCESS)
         {
-            USP_LOG_Error("%s: Failed to add response topic string", __FUNCTION__);
+            SaveMqttPublishErrMsg("%s: Failed to add response topic string", __FUNCTION__);
             err = USP_ERR_INTERNAL_ERROR;
             goto error;
         }
@@ -3838,16 +3872,16 @@ int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
     // Exit if property check failed
     if (mosquitto_property_check_all(PUBLISH, proplist) != MOSQ_ERR_SUCCESS)
     {
-        USP_LOG_Error("%s: property check failed.", __FUNCTION__);
+        SaveMqttPublishErrMsg("%s: property check failed.", __FUNCTION__);
         err = USP_ERR_INTERNAL_ERROR;
         goto error;
     }
 
-    // Exit if unableproperty check failed
+    // Exit if unable to publish the packet
     mosq_err = mosquitto_publish_v5(client->mosq, &msg->mid, msg->topic, msg->item.pbuf_len, msg->item.pbuf, msg->qos, false /* retain */, proplist);
     if (mosq_err != MOSQ_ERR_SUCCESS)
     {
-        USP_LOG_Error("%s: Failed to publish to v5 with error %d", __FUNCTION__, mosq_err);
+        SaveMqttPublishErrMsg("%s: Failed to publish to v5 with error %d (%s)", __FUNCTION__, mosq_err, mosquitto_strerror(mosq_err));
         err = USP_ERR_INTERNAL_ERROR;
         goto error;
     }
@@ -4663,6 +4697,29 @@ void RemoveExpiredMqttMessages(mqtt_client_t *client)
     }
 }
 
+/*********************************************************************//**
+**
+** SaveMqttPublishErrMsg
+**
+** Saves the cause of failure for SendQueueHead() in mqtt_publish_err_msg[]
+** so that is is available to put in an Error message that replaces the message that failed to send
+**
+** \param   fmt - printf style format
+**
+** \return  None
+**
+**************************************************************************/
+void SaveMqttPublishErrMsg(const char *fmt, ...)
+{
+    va_list ap;
 
+    // Print the message to the buffer, so that it can be used later
+    va_start(ap, fmt);
+    vsnprintf(mqtt_publish_err_msg, sizeof(mqtt_publish_err_msg), fmt, ap);
+    va_end(ap);
+
+    // Log to the log file too
+    USP_LOG_Error("%s", mqtt_publish_err_msg);
+}
 
 #endif